+++ browser/main/Detail/SnippetNoteDetail.js
@@ -17,7 +17,6 @@ import context from 'browser/lib/context'
 import ConfigManager from 'browser/main/lib/ConfigManager'
 import _ from 'lodash'
 import { findNoteTitle } from 'browser/lib/findNoteTitle'
-import AwsMobileAnalyticsConfig from 'browser/main/lib/AwsMobileAnalyticsConfig'
 import FullscreenButton from './FullscreenButton'
 import TrashButton from './TrashButton'
 import RestoreButton from './RestoreButton'
@@ -163,7 +162,6 @@ class SnippetNoteDetail extends React.Component {
         type: 'UPDATE_NOTE',
         note: note
       })
-      AwsMobileAnalyticsConfig.recordDynamicCustomEvent('EDIT_NOTE')
     })
   }
 
@@ -207,9 +205,6 @@ class SnippetNoteDetail extends React.Component {
 
   handleStarButtonClick(e) {
     const { note } = this.state
-    if (!note.isStarred)
-      AwsMobileAnalyticsConfig.recordDynamicCustomEvent('ADD_STAR')
-
     note.isStarred = !note.isStarred
 
     this.setState(
@@ -421,9 +416,6 @@ class SnippetNoteDetail extends React.Component {
     const mode = syntax != null ? syntax.name : null
     if (mode != null) {
       snippets[index].mode = mode
-      AwsMobileAnalyticsConfig.recordDynamicCustomEvent('SNIPPET_LANG', {
-        name: mode
-      })
     }
     this.setState(state => ({
       note: Object.assign(state.note, { snippets: snippets })
@@ -455,10 +447,6 @@ class SnippetNoteDetail extends React.Component {
           this.save()
         }
       )
-
-      AwsMobileAnalyticsConfig.recordDynamicCustomEvent('SELECT_LANG', {
-        name
-      })
     }
   }