+++ browser/main/Main.js
@@ -10,7 +10,6 @@ import Detail from './Detail'
 import dataApi from 'browser/main/lib/dataApi'
 import _ from 'lodash'
 import ConfigManager from 'browser/main/lib/ConfigManager'
-import mobileAnalytics from 'browser/main/lib/AwsMobileAnalyticsConfig'
 import eventEmitter from 'browser/main/lib/eventEmitter'
 import { store } from 'browser/main/store'
 import i18n from 'browser/lib/i18n'
@@ -27,10 +26,6 @@ const { remote } = electron
 class Main extends React.Component {
   constructor(props) {
     super(props)
-
-    if (process.env.NODE_ENV === 'production') {
-      mobileAnalytics.initAwsMobileAnalytics()
-    }
 
     const { config } = props